test(macros): compile mutable core inject paths
Move mutable identifier and destructuring injection coverage into the registered integration trybuild harness so route, use-inject, and route-registration expansions are compiled against framework dependencies.

Fixes #5773
